web: Refactor Sidebar component
