--- web/src/Sidebar.scss
@@ -1,55 +1,0 @@
-@import "constants";
-@import "animations";
-
-.Sidebar {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: $statusbar-height;
-  width: $sidebar-width;
-  background-color: $color-gray-darkest;
-  box-sizing: border-box;
-  overflow-y: auto;
-  transform: translateX(0%);
-  transition: transform ease $animation-timing;
-  font-size: $font-size;
-  display: flex;
-  flex-direction: column;
-  z-index: $z-sidebar;
-}
-.Sidebar.is-closed {
-  transform: translateX(calc(100% - #{$sidebar-collapsed-width}));
-}
-
-// Collapse/Expand
-.Sidebar-toggle {
-  background-color: $color-gray-darkest;
-  border: 0 none;
-  color: inherit;
-  font-size: inherit;
-  font-family: inherit;
-  display: flex;
-  align-items: center;
-  min-height: $sidebar-item;
-  margin: 0;
-  padding: 0;
-  @include button-text;
-  cursor: pointer;
-  transition-property: color, background-color;
-  transition-duration: $animation-timing;
-  transition-timing-function: ease;
-}
-.Sidebar-toggle:hover {
-  color: $color-blue-light;
-}
-.Sidebar-toggle > svg {
-  fill: $color-white;
-  width: $sidebar-collapsed-width;
-  transition: transform $animation-timing ease-in, fill $animation-timing ease;
-}
-.Sidebar-toggle:hover > svg {
-  fill: $color-blue-light;
-}
-.Sidebar.is-closed .Sidebar-toggle > svg {
-  transform: rotate(180deg);
-}
