+++ web/src/Sidebar.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { storiesOf } from "@storybook/react"
-import Sidebar, { SidebarItem } from "./Sidebar"
+import SidebarResources, { SidebarItem } from "./SidebarResources"
 import {
   oneResourceView,
   oneResourceNoAlerts,
@@ -13,6 +13,10 @@ import { ResourceStatus, ResourceView, TriggerMode } from "./types"
 type Resource = Proto.webviewResource
 let pathBuilder = new PathBuilder("localhost", "/")
 
+const fakeToggleAccountMenu = () => {
+  console.log("Toggled account menu")
+}
+
 function twoItemSidebar() {
   let items = twoResourceView().resources.map(
     (res: Resource) => new SidebarItem(res)
@@ -20,11 +24,9 @@ function twoItemSidebar() {
   items[0].name = "snapshot-frontend-binary-long-name"
   return (
     <MemoryRouter initialEntries={["/"]}>
-      <Sidebar
-        isClosed={false}
+      <SidebarResources
         items={items}
         selected=""
-        toggleSidebar={null}
         resourceView={ResourceView.Log}
         pathBuilder={pathBuilder}
       />
@@ -39,11 +41,9 @@ function twoItemSidebarClosed() {
   items[0].name = "snapshot-frontend-binary-long-name"
   return (
     <MemoryRouter initialEntries={["/"]}>
-      <Sidebar
-        isClosed={true}
+      <SidebarResources
         items={items}
         selected=""
-        toggleSidebar={null}
         resourceView={ResourceView.Log}
         pathBuilder={pathBuilder}
       />
@@ -61,11 +61,9 @@ function oneItemWithTrigger() {
   })
   return (
     <MemoryRouter initialEntries={["/"]}>
-      <Sidebar
-        isClosed={false}
+      <SidebarResources
         items={items}
         selected=""
-        toggleSidebar={null}
         resourceView={ResourceView.Log}
         pathBuilder={pathBuilder}
       />
@@ -86,11 +84,9 @@ function oneItemWithStatus(status: ResourceStatus) {
   let items = [item]
   return (
     <MemoryRouter initialEntries={["/"]}>
-      <Sidebar
-        isClosed={false}
+      <SidebarResources
         items={items}
         selected=""
-        toggleSidebar={null}
         resourceView={ResourceView.Log}
         pathBuilder={pathBuilder}
       />
