+++ web/src/style-helpers.ts
@@ -64,16 +64,16 @@ export enum Height {
   secondaryNavLower = unit * 0.8,
   secondaryNavOverlap = unit * -0.2,
   secondaryNavTwoLevel = unit * 1.8,
-  sidebarItem = unit * 1.4, // sync with constants.scss > $sidebar-item
-  resourceBar = unit * 1.5,
+  sidebarItem = unit * 1.4,
   statusbar = unit * 1.5,
 }
 export enum Width {
-  sidebar = unit * 10.5, // Sync with constants.scss > $sidebar-width
+  badge = unit * 0.6,
   secondaryNavItem = unit * 5,
   sidebarTriggerButton = unit * 1.4,
-  sidebarCollapsed = unit, // sync with constants.scss > $sidebar-collapsed-width
-  badge = unit * 0.6,
+  sidebar = unit * 10.5, // Sync with constants.scss > $sidebar-width
+  sidebarCollapsed = unit,
+  statusbar = unit * 1.5, // sync with constants.scss > $statusbar-height
   smallScreen = 1500,
 }
 
@@ -83,6 +83,7 @@ export const mixinHideOnSmallScreen = `
 }`
 
 export enum ZIndex {
+  Sidebar = 1000,
   HUDHeader = 500,
 }
